+++ JeffFerguson.Gepsio/Xml/Implementations/SystemXml/Node.cs
@@ -1,4 +1,5 @@
 ﻿using JeffFerguson.Gepsio.Xml.Interfaces;
+using System;
 using System.Xml;
 
 namespace JeffFerguson.Gepsio.Xml.Implementation.SystemXml
@@ -172,6 +173,42 @@ namespace JeffFerguson.Gepsio.Xml.Implementation.SystemXml
             if (OtherNode == null)
                 return false;
             return object.ReferenceEquals((this.ParentNode as Node).thisNode, (OtherNode.ParentNode as Node).thisNode);
+        }
+
+        /// <summary>
+        /// The value of the node typed to the data type specified in
+        /// the schema definition for the node. If no data type is available,
+        /// then a string representation is returned, in which case TypedValue
+        /// returns the same string as what the InnerText property returns.
+        /// </summary>
+        /// <param name="containingFragment">
+        /// The fragment containing the attributes.
+        /// </param>
+        /// <returns>
+        /// The value of the node typed to the data type specified in
+        /// the schema definition for the node. If no data type is available,
+        /// then a string representation is returned.
+        /// </returns>
+        public object GetTypedValue(XbrlFragment containingFragment)
+        {
+            throw new NotImplementedException();
+        }
+
+        /// <summary>
+        /// Compares the typed value of this node with the typed value of another node.
+        /// </summary>
+        /// <param name="otherNode">
+        /// The other node whose typed value is to be compared with this node's typed value.
+        /// </param>
+        /// <param name="containingFragment">
+        /// The fragment containing the attributes.
+        /// </param>
+        /// <returns>
+        /// True if the nodes have the same typed value; false otherwise.
+        /// </returns>
+        public bool TypedValueEquals(INode otherNode, XbrlFragment containingFragment)
+        {
+            throw new NotImplementedException();
         }
     }
 }